# swagger.yaml
openapi: 3.0.0
info:
  title: Barangay Census API
  version: 1.0.0
  description: >-
    An API that allows Barangays to Add people and and barangays and calculate
    the population automatically
  termsOfService: http://swagger.io/terms/
  contact:
    email: 20180015014@my.xu.edu.ph
  license:
    name: Node Js Licence
    url: https://github.com/nodejs/node/blob/main/LICENSE
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: APIKey
      in: header
...
paths:
...
  /users/register:
    post:
      summary: Register a new user.
      tags:
        - Users
      requestBody:
        description: User registration
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User registered successfully.
        '400':
          description: Client error.
        '500':
          description: Internal server error.
  /users/login:
    post:
      summary: Login with username and password.
      tags:
        - Users
      requestBody:
        description: User login
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful.
        '401':
          description: Invalid username or password.
        '500':
          description: Internal server error.
...